+++ Molten.Math/SinglePrecision/Vectors/SByte3.v3.cs
@@ -8,42 +8,42 @@ namespace Molten
            /// <summary>
         /// A unit <see cref="SByte3"/> designating up (0, 1, 0).
         /// </summary>
-        public static readonly SByte3 Up = new SByte3(0, 1, 0);
+        public static readonly SByte3 Up = new SByte3((sbyte)0, (sbyte)1, (sbyte)0);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating down (0, -1, 0).
         /// </summary>
-        public static readonly SByte3 Down = new SByte3(0, -1, 0);
+        public static readonly SByte3 Down = new SByte3((sbyte)0, -(sbyte)1, (sbyte)0);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating left (-1, 0, 0).
         /// </summary>
-        public static readonly SByte3 Left = new SByte3(-1, 0, 0);
+        public static readonly SByte3 Left = new SByte3(-(sbyte)1, (sbyte)0, (sbyte)0);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating right (1, 0, 0).
         /// </summary>
-        public static readonly SByte3 Right = new SByte3(1, 0, 0);
+        public static readonly SByte3 Right = new SByte3((sbyte)1, (sbyte)0, (sbyte)0);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating forward in a right-handed coordinate system (0, 0, -1).
         /// </summary>
-        public static readonly SByte3 ForwardRH = new SByte3(0, 0, -1);
+        public static readonly SByte3 ForwardRH = new SByte3((sbyte)0, (sbyte)0, -(sbyte)1);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating forward in a left-handed coordinate system (0, 0, 1).
         /// </summary>
-        public static readonly SByte3 ForwardLH = new SByte3(0, 0, 1);
+        public static readonly SByte3 ForwardLH = new SByte3((sbyte)0, (sbyte)0, (sbyte)1);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating backward in a right-handed coordinate system (0, 0, 1).
         /// </summary>
-        public static readonly SByte3 BackwardRH = new SByte3(0, 0, 1);
+        public static readonly SByte3 BackwardRH = new SByte3((sbyte)0, (sbyte)0, (sbyte)1);
 
         /// <summary>
         /// A unit <see cref="SByte3"/> designating backward in a left-handed coordinate system (0, 0, -1).
         /// </summary>
-        public static readonly SByte3 BackwardLH = new SByte3(0, 0, -1);
+        public static readonly SByte3 BackwardLH = new SByte3((sbyte)0, (sbyte)0, -(sbyte)1);
 
 #region Static Methods
         /// <summary>
